requirements_vision.txt
# Core verifiers dependencies
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0
datasets>=2.14.0
peft>=0.7.0
trl>=0.7.0

# Vision-language specific dependencies
Pillow>=10.0.0
opencv-python>=4.8.0
clip-by-openai>=1.0
torchvision>=0.15.0

# Training and evaluation
wandb>=0.16.0
vllm>=0.2.0
openai>=1.0.0
numpy>=1.24.0
scipy>=1.11.0

# Additional utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: for better performance
flash-attn>=2.5.0
xformers>=0.0.22

# Optional: for advanced image processing
scikit-image>=0.21.0
imageio>=2.31.0